+++ tests/conftest.py
@@ -31,24 +31,6 @@ def looker_client(record_mode) -> Iterable[LookerClient]:
         yield client
 
 
-@pytest.fixture(scope="session")
-def runner(record_mode) -> Iterable[Runner]:
-    with vcr.use_cassette(
-        "tests/cassettes/init_runner.yaml",
-        filter_post_data_parameters=["client_id", "client_secret"],
-        filter_headers=["Authorization"],
-        record_mode=record_mode,
-    ):
-        runner = Runner(
-            project="eye_exam",
-            branch="pytest",
-            base_url="https://spectacles.looker.com",
-            client_id=os.environ.get("LOOKER_CLIENT_ID", ""),
-            client_secret=os.environ.get("LOOKER_CLIENT_SECRET", ""),
-        )
-        yield runner
-
-
 @pytest.fixture
 def dimension():
     return Dimension(
